% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_EM_Lambda.R
\name{adaptative_EM_Lambda}
\alias{adaptative_EM_Lambda}
\title{EM procedure for sub tree}
\usage{
adaptative_EM_Lambda(sub)
}
\arguments{
\item{sub}{a Output of extract_tree}
}
\value{
Estimated proportion for the sub tree
}
\description{
EM procedure for sub tree
}
\examples{
\dontrun{

#using res for the Wavelet_screaming exemple


sub_analysis <- function(res, lev_res )
{
 sub <- extract_tree(res,lev_res=lev_res)
 my_pi <- adaptative_EM_Lambda(sub)
 out <-  adaptative_Lambda (my_pi, sub)
 return(out)
}


sub_analysis(res, 6)

}
}
