% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_Lambda.R
\name{adaptative_Lambda}
\alias{adaptative_Lambda}
\title{Likelihood ratio for sub tree}
\usage{
adaptative_Lambda(my_pi, sub)
}
\arguments{
\item{my_pi}{a vector of the proportion of association per level of resolution.}

\item{sub}{Output of extract_treet.}
}
\value{
Value of the likelihood on a sub tree.
}
\description{
internal function objective function for sub trees.
}
\examples{
\dontrun{

#using res for the Wavelet_screaming exemple


sub_analysis <- function(res, lev_res )
{
 sub <- extract_tree(res,lev_res=lev_res)
 my_pi <- adaptative_EM_Lambda(sub)
 out <-  adaptative_Lambda (my_pi, sub)
 return(out)
}


sub_analysis(res, 6)

}
}
