% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_Simu_Lambda_null.R
\name{adaptative_Simu_Lambda_null}
\alias{adaptative_Simu_Lambda_null}
\title{Simulation of Lambda null distribution on a sub tree}
\usage{
adaptative_Simu_Lambda_null(nsimu, lambda, lev_res, ncp, sub, thresh)
}
\arguments{
\item{nsimu}{number of simulation to perform.}

\item{lambda}{the lambda parameter of the null distribution of the Bayes Factor.}

\item{lev_res}{the maximum level of resolution needed.}

\item{ncp}{the lambda parameter of the null distribution of the Bayes Factor, if not specified set at 0.}
}
\description{
Simulation of Lambda null distribution.
}
\details{
Use the theoretical null distribution of the generated Bayes factor to perform simulation of null distriution of the test statistics of the Wavelet screaming procedure.
}
\references{
Quan Zhou and Yongtao Guan, On the Null Distribution of Bayes Factors in linear Regression, Journal of the American Statistical Association, 518, 2017.
}
