% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_p.R
\name{analytical_p}
\alias{analytical_p}
\title{Analytical pvalue for for loci with only one Bayes Factor over a threshold}
\usage{
analytical_p(sub, lambda, ncp, thresh)
}
\arguments{
\item{lambda}{the lambda parameter of the null distribution of the Bayes Factor.}

\item{ncp}{the lambda parameter of the null distribution of the Bayes Factor, if not specified set at 0.}

\item{nsimu}{number of simulation to perform.}

\item{lev_res}{the maximum level of resolution needed.}
}
\description{
Compute the analytical pvalue for for loci with only one Bayes Factor over a threshold
}
\details{
Use the theoretical null distribution of the generated Bayes factor to perform simulation of null distriution of the test statistics of the Wavelet screaming procedure.
}
