% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Compute Beta value from Bayesian linear regression}
\usage{
betas(y, x, confounder, sigma_b, all)
}
\arguments{
\item{y}{phenotype vector/variable of interest, has to be numeric.}

\item{x}{numerical, variable to regress on.}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{sigma_b}{the parameter of the NIG prior used for the Bayes Factor computation. We advised to set this value between 0.1 and 0.2}

\item{all}{logical, if set as TRUE return all the Beta value (including the ones form the confounding factors). If set as FALSE only return the estimate for x, set as FALSE if missing.}
}
\value{
A named vector. First position the estimated value of the Lambda statistics, then the proportion of association per level of resolution then the computed Bayes Factor per wavelet coefficient.
}
\description{
Compute Beta value from Bayesian linear regression
}
\details{
The Wavelet_screaming function computes the Likelihood ratio used for testing significance of a genetic region. In addition it computes
the porportion of wavelets coefficients associated by level of resolution, and the Bayes factor used for this estimation. All the details
of the computation can be found in our paper "Wavelet Screaming: a novel look to GWAS data"
}
\examples{
\dontrun{

x <- rnorm(1000)
y <- x+3
sigma_b <- 0.2
betas(y=y,x=x,sigma_b = sigma_b,all=TRUE)
}
}
