% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_Beta.R
\name{Cov_Beta}
\alias{Cov_Beta}
\title{Compute posterior variance Beta value from Bayesian linear regression}
\usage{
Cov_Beta(y, confounder, sigma_b, all)
}
\arguments{
\item{y}{phenotype vector/variable of interest, has to be numeric.}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{sigma_b}{the parameter of the NIG prior used for the Bayes Factor computation. We advised to set this value between 0.1 and 0.2}

\item{all}{logical, if set as TRUE return all the Beta value (including the ones form the confounding factors). If set as FALSE only return the estimate for x, set as FALSE if missing.}
}
\value{
A matrix variance covariance matrix
}
\description{
Compute posterior variance Beta value from Bayesian linear regression
}
\details{
The Wavelet_screaming performed reverse regression so variance for all posterior distribution are equal.
}
\examples{
\dontrun{

x <- rnorm(1000)
y <- x+3
sigma_b <- 0.2
Cov_Beta(y=y,sigma_b = sigma_b,all=TRUE)
}
}
