% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fine_mapping.R
\name{fine_map}
\alias{fine_map}
\title{Definition of the regions with unthresholded Betas.}
\usage{
fine_map(res, lev_res, thresh, start, end, chr)
}
\arguments{
\item{res}{Output of Wavelet_screening, without Betas.}

\item{lev_res}{the maximum level of resolution of the previous analysis.}

\item{thresh}{numeric, Bayes factor threshold to defined the fine mapping. If missing set as 1.}

\item{start}{numeric, start in base pair of the analyzed regions.}

\item{chr}{numeric, end in base pair of the analyzed regions.}
}
\description{
Fine mapping tool for output of the Wavelet screening function.
}
\details{
return a list of chr, start end position, that corresponds of the sub-regions defined by the dyadic decomposition of wavelet that is associated with a Bayes factor over the defined threshold.
}
