% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_lambda1.R
\name{get_lambda1}
\alias{get_lambda1}
\title{Compute lambda1 for BF null distribution}
\usage{
get_lambda1(Y, confounder, sigma_b)
}
\arguments{
\item{Y}{a vector of the variable of interest.}

\item{confounder}{the confounding matrix with same a number of line equal to the length of Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{sigma_b}{value of the prior used in the Wavelet screaming.}
}
\description{
Compute the lambda parameter of the Bayes factors null distribution.
}
\details{
Compute the lambda parameter of the Bayes factors distribution using its closed form which is provided in the paper of Quan Zhou and Yongtao Guan.
}
\examples{
\dontrun{
Y <- rnorm(n=1000)
sigma <-0.2
get_lambda1(Y=Y,sigma_b=sigma)
}
}
\references{
Quan Zhou and Yongtao Guan, On the Null Distribution of Bayes Factors in linear Regression, Journal of the American Statistical Association, 518, 2017.
}
\seealso{
\code{\link{get_ncp}}
}
