% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncp.R
\name{get_ncp}
\alias{get_ncp}
\title{Compute ncp for BF null distribution}
\usage{
get_ncp(Y, Counfounder, WCs, sigma_b)
}
\arguments{
\item{Y}{a vector of the variable of interest.}

\item{Counfounder}{the counfounding matrix with same a number of line equal to the length of Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{WCs}{the vector of a specific wavelet coefficient for each inidivual.}

\item{sigma_b}{value of the prior used in the Wavelet screaming.}
}
\description{
Compute non central parameter of the Bayes factors distribution.
}
\details{
Compute non central parameter of the Bayes factors distribution using its closed form which is provided in the paper of Quan Zhou and Yongtao Guan.
}
\examples{
\dontrun{
Y <- rnorm(n=1000)
WCs <- rnorm(n=1000)
sigma <-0.2
get_ncp(Y=Y,WCs=WCs,sigma_b=sigma)
}
}
\references{
Quan Zhou and Yongtao Guan, On the Null Distribution of Bayes Factors in linear Regression, Journal of the American Statistical Association, 518, 2017.
}
\seealso{
\code{\link{get_lambda1}}
}
