% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda_stat.R
\name{Lambda_stat}
\alias{Lambda_stat}
\title{Internal likelyhood ratio}
\usage{
Lambda_stat(my_pi, my_bayes)
}
\arguments{
\item{my_pi}{a vector of the proportion of association per level of resolution.}

\item{my_bayes}{a vector of bayes factor for all the wavelet transform component.}
}
\description{
internal function objective function.
}
