% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_EM_Lambda.R
\name{max_EM_Lambda}
\alias{max_EM_Lambda}
\title{Internal EM procedure}
\usage{
max_EM_Lambda(my_bayes, lev_res)
}
\arguments{
\item{my_bayes}{a vector of Bayes Factors of length 2^lev_res}

\item{lev_res}{the level of resolution in the wavelet transform}
}
\description{
Internal EM procedure
}
