% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbe_of_Wavelet_coefficients.R
\name{n_coef_wc}
\alias{n_coef_wc}
\title{Number of wavelet coefficients for a given level of resolution}
\usage{
n_coef_wc(lev_res)
}
\arguments{
\item{lev_res}{the maximum level of resolution needed}
}
\description{
Compute the number of wavelet coefficients up to a certain levl of resolution
}
\examples{
\dontrun{
n_coef_wc(7)
}
}
