% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_sub.R
\name{prop_sub}
\alias{prop_sub}
\title{Compute the relative size of a sub region.}
\usage{
prop_sub(sub)
}
\arguments{
\item{sub}{Output of extract_tree.}
}
\value{
A proportion, used for multiple testing correction for the region
}
\description{
Compute the relative size of a sub region.
}
