% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Lambda_null_distribution.R
\name{Simu_Lambda_null}
\alias{Simu_Lambda_null}
\title{Simulation of Lambda null distribution}
\usage{
Simu_Lambda_null(nsimu, lambda, lev_res, ncp)
}
\arguments{
\item{nsimu}{number of simulation to perform.}

\item{lambda}{the lambda parameter of the null distribution of the Bayes Factor.}

\item{lev_res}{the maximum level of resolution needed.}

\item{ncp}{the lambda parameter of the null distribution of the Bayes Factor, if not specified set at 0.}
}
\description{
Simulation of Lambda null distribution.
}
\details{
Use the theoretical null distribution of the generated Bayes factor to perform simulation of null distriution of the test statistics of the Wavelet screaming procedure.
}
\examples{
\dontrun{
set.seed(1)
Simu_Lambda_null(100,0.7,2)
}
}
\references{
Quan Zhou and Yongtao Guan, On the Null Distribution of Bayes Factors in linear Regression, Journal of the American Statistical Association, 518, 2017.
}
\seealso{
\code{\link{Wavelet_screaming}}
}
