% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_null.R
\name{Simu_null}
\alias{Simu_null}
\title{Simulation of the null statsitics}
\usage{
Simu_null(Y, confounder, lev_res, emp_cov, size, sigma_b, print = TRUE)
}
\arguments{
\item{Y}{a vector of numeric values used in in the wavelet screaming function for association.}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{lev_res}{the level of resolution in the wavelet transform}

\item{emp_cov}{Emprical covariance matrix of the beta values. Can be computed using several results of the wavelet screaming using the betas values for different loci. If missing the function compute an approximation of the covariance matrix, this leads to a lost of power and a more conservative test statitics.}

\item{size}{number of simulation to be performed}

\item{sigma_b}{the parameter of the NIG prior used for the Betas computation.}

\item{print}{logical parameter set as TRUE, if TRUE send message when 10\% of the simulations has been completed.}
}
\value{
The simulation under the null of the two test statistics used to build the final test (i.e L_h and min(ph,pv))
}
\description{
Simulation of the null statsitics
}
\examples{
\dontrun{
Y <- rnorm(4000)
Sim <- Simu_null(Y=Y,lev_res = 6,sigma_b = 0.2,size=10000)
}
}
