% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_null_proxy.R
\name{Simu_null_proxy}
\alias{Simu_null_proxy}
\title{Simulation of the null statistics}
\usage{
Simu_null_proxy(Y, confounder, lev_res, size, sigma_b, print = TRUE)
}
\arguments{
\item{Y}{a vector of numeric values used in the wavelet screaming function for association (a phenotype or simulated phenotype with the same distribution).}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included if missing will generate an intercept matrix.}

\item{lev_res}{the level of resolution in the wavelet transform}

\item{size}{number of simulation to be performed}

\item{sigma_b}{the parameter of the NIG prior used for the Betas computation.}

\item{print}{logical parameter set as TRUE, if TRUE sends a message when 10\% of the simulations have been completed.}
}
\value{
The simulation under the null of the two test statistics used to build the final test (i.e., L_h and min(ph,pv))
}
\description{
Simulation of the null statistics
}
\examples{
\dontrun{
Y <- rnorm(4000)
Sim <- Simu_null_proxy(Y=Y,lev_res = 6,sigma_b = 0.2,size=10000)
}
}
