% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelet_processing.R
\name{wavproc}
\alias{wavproc}
\title{Internal wavelet processing}
\usage{
wavproc(Time01, y, lev_res)
}
\arguments{
\item{Time01}{rescalled measurment positions (ordered)}

\item{y}{the signal}

\item{lev_res}{the maximum level of resolution needed}
}
\description{
Internal function perform a wavelet transform and thresholding
}
