% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree.R
\name{extract_tree}
\alias{extract_tree}
\title{Extract sub result from result of the wavelet_screaming}
\usage{
extract_tree(res, lev_res, thresh)
}
\arguments{
\item{res}{Output of Wavelet_screaming.}

\item{lev_res}{the maximum level of resolution needed, has to be less or equal to the request level of resolution in the Wavelet_screaming.}

\item{thresh}{Minimal value of the Bayes Factor to  defined the a sub region, if missing set as 1.}
}
\value{
A vector correpsonding of the sub tree for the zoomed analysis.
}
\description{
Function to performed zoomed analysis of the wavelet screaming output
}
\examples{
\dontrun{

#using res for the Wavelet_screaming exemple


sub_analysis <- function(res, lev_res )
{
 sub <- extract_tree(res,lev_res=lev_res)
 my_pi <- adaptative_EM_Lambda(sub)
 out <-  adaptative_Lambda (my_pi, sub)
 return(out)
}


sub_analysis(res, 6)

}
}
