% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile_transform.R
\name{Quantile_transform}
\alias{Quantile_transform}
\title{Perform a quantile transform}
\usage{
Quantile_transform(x)
}
\arguments{
\item{x}{a vector to be quantile transform}
}
\description{
Perform a quantile transform with ties taken at random
}
\examples{
\dontrun{
x <- rnorm(n=1000)
Quantile_transform(x)
}
}
