% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu_null_emp.R
\name{Simu_null_emp}
\alias{Simu_null_emp}
\title{Simulation of the null statistics}
\usage{
Simu_null_emp(emp_cov, smp_size, lev_res, size, sigma_b, print = TRUE)
}
\arguments{
\item{emp_cov}{Empirical covariance matrix of the beta values. It can be computed using several results of the wavelet screaming using the betas values for different loci. If missing the function computes an approximation of the covariance matrix, this leads to a loss of power and a more conservative test statistics.}

\item{smp_size}{Sample size from the main run of Wavelet Screaming}

\item{lev_res}{the level of resolution in the wavelet transform}

\item{size}{number of simulation to be performed}

\item{sigma_b}{the parameter of the NIG prior used for the Betas computation.}

\item{print}{logical parameter set as TRUE, if TRUE sends a message when 10\% of the simulations have been completed.}
}
\value{
The simulation under the null of the two test statistics used to build the final test (i.e., L_h and min(ph,pv))
}
\description{
Simulation of the null statistics
}
