% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelet_screaming.R
\name{Wavelet_screaming}
\alias{Wavelet_screaming}
\title{Main function to perform wavelet screaming}
\usage{
Wavelet_screaming(
  Y,
  loci,
  bp,
  confounder,
  lev_res,
  sigma_b,
  coeftype = "d",
  para = FALSE,
  BF = FALSE
)
}
\arguments{
\item{Y}{phenotype vector has to be numeric. For case-control  data, code it as 0 and 1. Multiple label phenotypes, e.g., ABO blood groups, will be implemented in the next version.}

\item{loci}{genotype matrix (either data.frame or numeric matrix). Lines=SNPs in increasing order in terms of base pair, columns=individuals. No missing values allowed.}

\item{bp}{vector of the base pairs positions. It has to be in the same order and length than the locus line order/length.}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included if missing will generate an intercept matrix.}

\item{lev_res}{the maximum level of resolution needed.}

\item{sigma_b}{the parameter of the NIG prior used for the Beta computation. We advised setting this value between 0.1 and 0.2}

\item{coeftype}{type of wavelet coefficient used for the screening.}

\item{para}{logical parameter for parallelization, if not specified, set at FALSE by default.}

\item{BF}{logical parameter for obtainning the Bayes Factor of the wavelet regression. If not specified, set at FALSEby default.}
}
\value{

}
\description{
Perform a wavelet screening of a locus for a given phenotype and a specified level of resolution
}
\details{
The Wavelet_screaming function computes the likelihood ratio used for testing the significance of a genetic region. In addition, it computes
the proportion of wavelet coefficients associated by the level of resolution and the Beta used for this estimation. All the details
of the computation can be found in our paper, preliminarily titled "Wavelet Screaming: a novel look to GWAS data.".
}
