% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelet_screaming.R
\name{Wavelet_screaming}
\alias{Wavelet_screaming}
\title{Main function to perform wavelet screaming}
\usage{
Wavelet_screaming(Y, loci, bp, confounder, lev_res, sigma_b,
  coeftype = "d", para = FALSE)
}
\arguments{
\item{Y}{phenotype vector, has to be numeric. For case control code it as 0 and 1. Multiple label phenotype will be implemented in the next version}

\item{loci}{genotype matrix (either data.frame or numeric matrix). Lines=SNPs in increasing order in term of base pair, columns=individuals. No missing values allowed.}

\item{bp}{vector of the base pairs positions. It has to be in the same order and length than the loci line order/length.}

\item{confounder}{the confounding matrix with the same sample order as Y. The intercept should not be included, if missing will generate a intercept matrix.}

\item{lev_res}{the maximum level of resolution needed}

\item{sigma_b}{the parameter of the NIG prior used for the Bayes Factor computation. We advised to set this value between 0.1 and 0.2}

\item{coeftype}{type of wavelet coefficient used for the screening. By default set as "d" difference, "c" can be used if you prefere to work in term of amount of variants instead in disrepency within sub loci.}

\item{para}{logical parameter for parallelisation, if not specified set at FALSE.}
}
\value{
A named vector. First position the estimated value of the Lambda statistics, then the proportion of association per level of resolution then the computed Bayes Factor per wavelet coefficient.
}
\description{
Perform a wavelet screening of a loci for a given phenotype and a specified level of resolution
}
\details{
The Wavelet_screaming function computes the Likelihood ratio used for testing significance of a genetic region. In addition it computes
the porportion of wavelets coefficients associated by level of resolution, and the Bayes factor used for this estimation. All the details
of the computation can be found in our paper "Wavelet Screaming: a novel look to GWAS data"
}
\examples{
\dontrun{


set.seed(666)
#########################################
#Generate a randomly sample loci size=1Mb
#########################################

#5000 Randomly choosen bp
my_bp <- sort(sample(1:1000000, size=5000,replace = FALSE))
#############################
#Three different bump signals
#############################
my_functions <-data.frame(f0 = c(rep(0,400000),rep(0,200000),rep(0,400000)),
                         f1 = c(rep(0,400000),rep(1,200000),rep(0,400000)) ,
                         f2=c(rep(0,400000),rep(2,200000),rep(0,400000)))


library(gridExtra)
###########################
#Minor allele frequency 30\%
###########################
MAF=0.3
sampl_schem <- c((1-MAF)^2,2*MAF*(1-MAF),MAF^2)
#######################################
#sampling at Hardy Weinberg equilibrium
#######################################
#Assigning class

#sample size =4000
n_size=4000
type_fn <-sample(0:2,replace = TRUE,size=n_size,  prob=  sampl_schem  )


genotype <-  matrix(my_functions[my_bp,2 ], ncol=1 ) \%*\%t(matrix(type_fn,ncol=1))
#dim(genotype)= nSNP, nind

###############################################################
#Generate a phenotype with variance explained by genotype  0.5\%
###############################################################
varexp=0.005
var_noise <- (1-varexp)*var(sample(0:2,replace = TRUE,size=10000,
                                  prob=sampl_schem ))/varexp
Y <-  rnorm(n=n_size,sd=sqrt(var_noise)) +type_fn
df <- data.frame(y=Y,genotype =factor(type_fn))
P1 <- ggplot(df,aes(y=y,x=genotype))+
 geom_boxplot()+
 xlab("Type of genotype")+
 theme(axis.text=element_text(size=12),
       axis.title=element_text(size=14,face="bold"))+
 ylab("Simulated Phenotype")+
 theme_bw()+
 ggtitle("Variation of the phenotype\\ndepending of the genotype, \\nVariance explained =0.5\%")

df <- data.frame(bp= rep(my_bp,3),y=c(my_functions[my_bp,1],my_functions[my_bp,2],my_functions[my_bp,3]),
                mycol = factor(c(rep("f0",length(my_bp)),rep("f1",length(my_bp)),rep("f2",length(my_bp))) ) )

P2 <- ggplot(df,aes(y=y,x=bp,color=mycol))+
 geom_point(size=1)+
 xlab("Base pair")+
 ylab("Number of variants")+
 theme_bw()+
 theme(legend.title=element_blank())+
 ggtitle("Three different kind of genotype signal")

grid.arrange(P1,P2,ncol=2)

##################
#Wavelet screaming
##################
res <- Wavelet_screaming( Y,loci=genotype,bp=my_bp,
                         lev_res=6,sigma_b = 0.2)
# or:
genotype_df <- as.data.frame(genotype)
res <- Wavelet_screaming( Y,loci=genotype_df,bp=my_bp,
                         lev_res=6,sigma_b = 0.2)
#value of the test statistic
res["Lambda"]
#############
#Significance
#############

#Estimation of the Bayes factor lambda_1 distribution parameter
#Take a bit of time
lambda <- get_lambda1(Y,sigma_b = 0.2)
#Simulation of the test statistics under the null distribution of
# the Bayes factor
Sim_gam <- Simu_Lambda_null(nsimu=10000, lambda=lambda,lev_res = 6)
val <- res["Lambda"]

#Via Simulation

pval <-c(length(Sim_gam[which(Sim_gam>val)])+1)/(length(Sim_gam)+1)
pval

#Via EVT
#Should be preferred for smaller values of Lambda

library(fExtremes)
x <-  Sim_gam
#Selecting the threshold
thresh <- min(  max (c( 2.5*mean( Sim_Gamma) , quantile(Sim_Gamma,0.98)) ), quantile(Sim_Gamma,0.97))
z = gpdFit(x, u = thresh , type = "mle")
z
pval <- 1-fExtremes::pgpd(q=res["Lambda"], xi=z@fit$par.ests["xi"], mu=z@parameter$u, beta=z@fit$par.ests["beta"])
pval


##############
#Visualisation
##############
bp <- c(min(my_bp),max(my_bp))
plot_WS(res=res,bp=bp,lev_res=6)


}
}
